+++ src/Symfony/Bundle/SecurityBundle/Tests/DependencyInjection/CompleteConfigurationTest.php
@@ -81,6 +81,11 @@ abstract class CompleteConfigurationTest extends TestCase
             $configs[] = array_values($configDef->getArguments());
         }
 
+        // the IDs of the services are case sensitive or insensitive depending on
+        // the Symfony version. Transform them to lowercase to simplify tests.
+        $configs[0][2] = strtolower($configs[0][2]);
+        $configs[2][2] = strtolower($configs[2][2]);
+
         $this->assertEquals(array(
             array(
                 'simple',