+++ src/Symfony/Component/DependencyInjection/Dumper/YamlDumper.php
@@ -110,6 +110,14 @@ class YamlDumper extends Dumper
             $code .= "        autowire: true\n";
         }
 
+        if ($definition->isAutoconfigured()) {
+            $code .= "        autoconfigure: true\n";
+        }
+
+        if ($definition->isAbstract()) {
+            $code .= "        abstract: true\n";
+        }
+
         if ($definition->isLazy()) {
             $code .= "        lazy: true\n";
         }